requirements.txt
fastapi>=0.111.0
supabase>=2.4.6
python-dotenv>=1.0.0
loguru>=0.7.0
pydantic>=2.7.0
